#108 Generating an OpenAPI view with Kotlin
In a multi-module gradle project that used Kotlin with a kotlin script file instead of a groovy script file, I was able to successfully generate the Open API views when setting the `org.gradle.jvmargs` property. This allowed me to not have to use the environment variable approach to generate the views correctly. When trying to use the `kapt` block in my `build.gradle.kts`, I could not successfully generate the views.