Replace adhoc array type helper (needs micronaut 2.2+)
Since micronaut-core 2.2.0, `ClassElement` is equipped with a
`fromArray` method we're better to use in lieu of the temporary helper
that was added (#368) in order to fix the generation of multi-
dimensional arrays (#367). This removes the risk that wrong array
dimensions get inadvertently used.

This partially reverts commit a4524550d45c36d78061ab392bdf41d274188cf6.