All not nullable parameters in a constructor are required in OpenAPI
Fixes #409

According to https://swagger.io/docs/specification/data-models/data-types/#required
all object properties are optional by default, so if a fields is not
nullable, it should be declared as "required" in the OpenAPI spec file.

In case the `@Schema` annotation exists on the field, it takes
precendece because the user decided to use it. It's up to the user to
make sure that the annotations in the field and in the constructor are
the same so they make sense.