Duplicate managed-swagger depenedency for core compatibility
We used to declare managed-swagger in the core catalog.  We want to remove this for ease of
maintenance, but this means we would break compatibility as here it's managed-swagger-core.

This PR adds a temporary duplication of the catalog to keep managed-swagger in the catalog here

This can be removed for 4.0.0