Make it possible to use a custom code generator
This commit extends the builder API in order to be able
to use custom code generators. This is required whenever the
standard code generators are not enough. In this case, the
user would have to provide not only the code generator, but
also implement a builder to configure the options specific
for that code generator.